+++ miamoremustachio/ToDo/modules/helpers/checking.js
